auth-service/requirements.txt
fastapi
uvicorn[standard]
pydantic[email,url]
python-dotenv
httpx
PyJWT
google-cloud-firestore
google-cloud-secret-manager
SQLAlchemy
asyncpg
google-cloud-kms
black
flake8
isort
pydantic-settings